vector: merge 'vector-type.h' into 'vector-priv.h'
It's not good way to maintain source code that type definition is
separated from others, for instance, one file is existed for only
define a type.

Instead of it, merge the type into private area. So we don't need
to move back and forth to figure out how the type defined.