vector-priv: improve 'insert_at()'/'remove_at()' performance
These two functions used a simple loop algorithm to adjust the
position of each elements when a new element pushed in the middle
of them. In that case, it moved each elements one step forward/backward
from where themselves were. It is totally bad when they called inside
a loop, for instance, 'for'.

Use 'memmove' instead of the simple algorithm. We give the responsibility
for the task to libc. Which means that we don't care how it implements
'memmove' function and which algorithm used. But we believe that it has
better algorithm than ours.

glibc, uclibc, musl and etc. There are more than described here.