vector: add/change APIs and a small news
* vec_add
- dzf_vec_add_head
- dzf_vec_add_tail

* vec_remove
- dzf_vec_rmv_at
- dzf_vec_rmv_head
- dzf_vec_rmv_tail

Rename 'remove' to 'rmv'. The former one was too long and has bad UX.

* News
__dzf_vec_insert_at() and __dzf_vec_remove_at() were defined as
'statement'. In that case, we couldn't use them in 'expression' syntax.
For instance, the below syntax is wrong. Compile error.

   for (int i = 0; i < 5; dzf_vec_add_back(&myspace, i)) /* loop 5 times */;
			  ^
			  ` statement couldn't be here

But since a patch[1] was committed, it is as a 'expression' syntax and
the above syntax totally fine. Compilers guarantee it.

[1]: Commit: a0b24bb7d3f7a2d86e47940ab5edf53fd109be01